śmiga, ale do wytrenowania

+++ src/hand_gesture_recognition/training.py
@@ -6,19 +6,34 @@ from sklearn.model_selection import train_test_split
 from sklearn.preprocessing import LabelEncoder
 
 
-def load_data(csv_file):
+def normalize_data(data, width=960, height=540):
+    normalized_data = []
+
+    for i, row in data.iterrows():
+        lm_list = []
+
+        for j in range(1, 22):
+            x = row[f'x{j}'] / width
+            y = row[f'y{j}'] / height
+            z = row[f'z{j}']
+            lm_list.extend([x, y, z])
+
+        normalized_data.append(lm_list)
+
+    return np.array(normalized_data)
+
+
+def load_and_preprocess_data(csv_file):
     data = pd.read_csv(csv_file)
     data.columns = data.columns.str.strip()
 
     data = data.drop(columns=['empty'], errors='ignore')
 
-    X = data.iloc[:, 1:-1].values
-    y = data['gesture_name'].values
+    X = normalize_data(data)
 
+    y = data['gesture_name'].values
     label_encoder = LabelEncoder()
     y = label_encoder.fit_transform(y)
-
-    X = X / np.max(X, axis=0)
 
     return X, y, label_encoder
 
@@ -39,24 +54,26 @@ def build_model(input_shape, num_classes):
 
 def main():
     csv_file = os.path.join("..", "..", "gestures", "data.csv")
-    X, y, label_encoder = load_data(csv_file)
+
+    if not os.path.exists(csv_file):
+        raise FileNotFoundError(f"Ścieżka {csv_file} jest nie poprawna.")
+
+    X, y, label_encoder = load_and_preprocess_data(csv_file)
 
     X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
 
+    input_shape = X_train.shape[1]
     num_classes = len(np.unique(y))
-    model = build_model(X_train.shape[1], num_classes)
+    model = build_model(input_shape, num_classes)
 
-    # Train the model
-    model.fit(X_train, y_train, epochs=50, batch_size=32, validation_split=0.2) # nauka tylko 50 linijek, zwiększyć potem
+    # amount of epochs to change
+    history = model.fit(X_train, y_train, epochs=100, batch_size=32, validation_split=0.2)
 
     loss, accuracy = model.evaluate(X_test, y_test)
     print(f"Test Accuracy: {accuracy:.2f}")
 
     model.save("gesture_recognition_model.keras")
-    print("Model saved as gesture_recognition_model.h5")
-
     np.save("label_encoder_classes.npy", label_encoder.classes_)
-    print("Label encoder classes saved as label_encoder_classes.npy")
 
 
 if __name__ == "__main__":
